Melhorias

+++ tipos_produto/remove.php
@@ -8,5 +8,13 @@
 
 require_once __DIR__ ."/../models/TipoProduto.php";
 $tipoProduto = new TipoProduto();
+require_once __DIR__ ."/../models/Produto.php";
+$produto = new Produto();
 
-header('Location: /tipos_produto/index.php?success='. $tipoProduto->delete($_GET['id']));
+$res = $produto->getAll($_GET['id']);
+
+if(!empty($res)){
+    header('Location: /tipos_produto/index.php?success=0');
+}else{
+    header('Location: /tipos_produto/index.php?success='. $tipoProduto->delete($_GET['id']));
+}
